+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * stepIgnore and stepIgnoreAll now return the function/table passed in, for convenience
 * Objects printed with `__DebugAdapter.print(obj)` will be expandable in debug console until the next break-and-continue or until the current lua stack ends
 * Pause button able to break into long running code
+* General code cleanup
+* Optional arguments `upStack` and `category` on `__DebugAdapter.print()`
 
 ## 0.18.36